+++ src/BrefServiceProvider.php
@@ -2,6 +2,8 @@
 
 namespace CacheWerk\BrefLaravelBridge;
 
+use CacheWerk\BrefLaravelBridge\Queue\QueueHandler;
+
 use Illuminate\Log\LogManager;
 
 use Illuminate\Support\Facades\Config;
@@ -58,6 +60,10 @@ class BrefServiceProvider extends ServiceProvider
         Config::set('queue.connections.sqs.secret');
         Config::set('queue.connections.sqs.token', env('AWS_SESSION_TOKEN'));
         Config::set('queue.connections.sqs.prefix', env('SQS_PREFIX', "https://sqs.{$region}.amazonaws.com/{$account}"));
+
+        $this->app->when(QueueHandler::class)
+            ->needs('$connection')
+            ->giveConfig('queue.default');
     }
 
     /**
@@ -77,10 +83,6 @@ class BrefServiceProvider extends ServiceProvider
             $this->publishes([
                 __DIR__ . '/../config/bref.php' => config_path('bref.php'),
             ], 'bref-config');
-
-            $this->publishes([
-                __DIR__ . '/../stubs/runtime.php' => base_path('php/runtime.php'),
-            ], 'bref-runtime');
         }
 
         $dispatcher->listen(